Squashed some bugs, plus general cleanup of Code

+++ localisationExample/src/Particle.h
@@ -23,7 +23,7 @@ class Particle
     // Propagate the sample based on the received odomotry information
     void propagateSample(double forwardMotion, double angleMotion, double proc_noise[2]);
     // Compute the likelihood of the particle based on the received measurement information
-    double computeLikelihood(measurementList measurement, World world, double meas_noise[2]);   
+    double long computeLikelihood(measurementList measurement, World world, double meas_noise[2]);   
     
     // Internal functions to compute uniform or gaussian samples
     double _get_noise_sample_unfiorm(double minval, double maxval);
@@ -37,7 +37,7 @@ class Particle
     double _y;
     double _theta;
 
-    double _weight;
+    double long _weight;
 
     std::default_random_engine* _generatorPtr;
 };
